fix: prep for other precompiles
